ENYO-3498: changes of ares-generate
ENYO-3498: added filesubstitution field in ide-plugin.json. this is a file list for substitution of file contents in template

ENYO-3498: used @ID@ keyword in template. this will be replaced with appID

ENYO-3498: Service ID & Name are sited as @ID@ in webos-service template. this will be replaced with appID.

Enyo-DCO-1.1-Signed-off-by Junil Kim <logyourself@gmail.com>